Rearranged files to clearly identify types
This commit rearranges and renames files so that it is clear what each
is. They are organized by type (controller or router) and segmented
minimally to make things simple going forward. The router in the root of
the api handles pulling in child routers so that server.js only needs to
pull in a single router for everything under /api.

The new file structure looks like this:

```
$ tree ./src
./src
├── api
│   ├── controllers
│   │   ├── errorController.js
│   │   └── pingController.js
│   ├── routes
│   │   └── index.js
│   └── v1
│       ├── controllers
│       │   └── pwr
│       │       ├── statsController.js
│       │       └── versionController.js
│       └── routes
│           └── pwr.js
├── server.js
└── util
    └── validation.js
```

This structure was heavily inspired by
https://developer.mozilla.org/en-US/docs/Learn/Server-side/Express_Nodejs/routes